copy rules and cookies responsibilities on to cookie_middleware
