+++ lib/sitehub/forward_proxy.rb
@@ -1,6 +1,5 @@
 # rubocop:disable Metrics/ParameterLists
-require 'sitehub/http_headers'
-require 'sitehub/string_sanitiser'
+require 'sitehub/request'
 require 'sitehub/request_mapping'
 require 'sitehub/rules'
 require 'sitehub/resolver'
@@ -15,58 +14,10 @@ require 'sitehub/constants'
 # 5. Expect header (optional)
 
 class SiteHub
-  class Request
-    include StringSanitiser, Constants, HttpHeaders
-
-    attr_reader :env, :rack_request
-    extend Forwardable
-
-    def_delegator :@rack_request, :params
-    def_delegator :@rack_request, :url
-    def_delegator :@rack_request, :path
-
-    def initialize(env)
-      @rack_request = Rack::Request.new(env)
-      @env = filter_http_headers(extract_http_headers_from_rack_env(env))
-    end
-
-    def request_method
-      @request_method ||= rack_request.request_method.downcase.to_sym
-    end
-
-    def body
-      @body ||= rack_request.body.read
-    end
-
-    def headers
-      @env.tap do |headers|
-        # x-forwarded-for
-        headers[X_FORWARDED_FOR_HEADER] = x_forwarded_for
-
-        # x-forwarded-host
-        headers[X_FORWARDED_HOST_HEADER] = x_forwarded_host
-      end
-    end
-
-    def remote_address
-      rack_request.env[RackHttpHeaderKeys::REMOTE_ADDRESS_ENV_KEY]
-    end
-
-    def x_forwarded_host
-      split(env[HttpHeaderKeys::X_FORWARDED_HOST_HEADER])
-        .push(env[HttpHeaderKeys::HOST_HEADER])
-        .join(COMMA)
-    end
-
-    def x_forwarded_for
-      split(env[HttpHeaderKeys::X_FORWARDED_FOR_HEADER]).push(remote_address).join(COMMA)
-    end
-  end
-
   class ForwardProxy
     ERROR_RESPONSE = Rack::Response.new(['error'], 500, {})
 
-    include HttpHeaders, Rules, Resolver, Constants, StringSanitiser
+    include Rules, Resolver, Constants
 
     attr_reader :url, :id, :mapped_path, :http_client, :sitehub_cookie_path, :sitehub_cookie_name
